Implement AOS and LOS signaling on rigctl interface.
This commit implements basic AOS and LOS signaling on the rigctl
interface by sending the strings "AOS" and "LOS" over the network
interface. This signal was introduced as a gqrx extension to the
rigctld protocol.

I have also prepared to make this setting configurable for each radio,
but it has not been fully implemented.