Switch to using complete URLs for TLE update
This changes from storing one server and a number of files to storing complete
URLs, allowing to use more than one server. An attempt has been made to import
existing config records so that existing users shouldn't notice any difference.

New users will only have the single AMSAT source we are switching to.

On the UI side we switch to using a list for the URLs instead of direct display
of the ";" separated list.