fixed Bug in JSON transponder parser
The behaviour of `sprintf()` is undefined when using the output buffer
as an argument. This may lead to parsers errors like that described in
issue #39. This commit shifts the preparsing of the transponder JSON
file entirely to nxjson, thus further avoiding a possible stack overflow
vulnerability through the `jsn_object` buffer.

Signed-off-by: Mario Haustein <mario.haustein@hrz.tu-chemnitz.de>