Catch negative result returned by ftell()
Upon successful completion, ftell() returns the current offset.
Otherwise, -1 is returned and errno is set to indicate the error.
The length indicator mfplen is changed to size_t to fit the fread()
prototype and for better portability. To hold the interim result
returned by ftell() a new placeholder long flen has been inserted.
The call of fread() is protected from being used with a negative
argument.

See: CID #1462972