src/main: Fix GdkMonitor acquisition failure
On a Wayland system, we can't get the primary monitor due to protocol
shortage[1].  This caused gpredict to print a critical message like:

  Gdk-CRITICAL **: 02:03:59.960: gdk_monitor_get_workarea: assertion
  'GDK_IS_MONITOR (monitor)' failed

Because it failed get the primary monitor, `w` and `h` were just plain
garbage.

Instead of trying to get the primary monitor, this commit gets the
monitor the window is on.  This commit has been tested both on a
Wayland and a X11 session.

https://gitlab.gnome.org/GNOME/gtk/-/issues/1028

Signed-off-by: Yasushi SHOJI <yasushi.shoji@gmail.com>